requirements.txt
langchain
langchain-core
langchain-community
langchain-qdrant
qdrant-client
google-generativeai
python-dotenv
pypdf